+++ ckks/bootstrapping/bootstrapping_test.go
@@ -27,7 +27,7 @@ func ParamsToString(params ckks.Parameters, opname string) string {
 		params.LogQP(),
 		params.MaxLevel()+1,
 		params.PCount(),
-		params.DecompRNS(params.QCount()-1, params.PCount()-1))
+		params.DecompRNS(params.MaxLevelQ(), params.MaxLevelP()))
 }
 
 func TestBootstrapParametersMarshalling(t *testing.T) {