+++ rlwe/keys.go
@@ -95,7 +95,7 @@ func NewRotationKeySet(params Parameters, galoisElement []uint64) (rotKey *Rotat
 	rotKey = new(RotationKeySet)
 	rotKey.Keys = make(map[uint64]*SwitchingKey, len(galoisElement))
 	for _, galEl := range galoisElement {
-		rotKey.Keys[galEl] = NewSwitchingKey(params, params.QCount()-1, params.PCount()-1)
+		rotKey.Keys[galEl] = NewSwitchingKey(params, params.MaxLevelQ(), params.MaxLevelP())
 	}
 	return
 }
@@ -136,7 +136,7 @@ func NewRelinearizationKey(params Parameters, maxRelinDegree int) (evakey *Relin
 	evakey = new(RelinearizationKey)
 	evakey.Keys = make([]*SwitchingKey, maxRelinDegree)
 	for d := 0; d < maxRelinDegree; d++ {
-		evakey.Keys[d] = NewSwitchingKey(params, params.QCount()-1, params.PCount()-1)
+		evakey.Keys[d] = NewSwitchingKey(params, params.MaxLevelQ(), params.MaxLevelP())
 	}
 
 	return